add dji com fibHeap. Tudo compilando, mas sem testes

+++ codigo/fibHeap.h
@@ -21,7 +21,7 @@ void fhLink(tFH *fh, tNodeFH *y, tNodeFH *x);
 
 void fhConsolidate(tFH *fh);
 
-void fhDecreaseKey(tFH *fh, tNodeFH *node, int newKey);
+void fhDecreaseKey(tFH *fh, tNodeFH *node, float newKey);
 
 bool fhIsEmpty(tFH *fh);
